requirements.txt
streamlit
deepface
opencv-python-headless
tf-keras
firebase-admin
pillow
requests
numpy